Move RH to indico.web.rh
