+++ indico/modules/bootstrap/controllers.py
@@ -26,7 +26,6 @@ import indico
 from indico.core.config import config
 from indico.core.db import db
 from indico.core.db.sqlalchemy.util.queries import get_postgres_version
-from indico.legacy.webinterface.rh.base import RH
 from indico.modules.auth import Identity, login_user
 from indico.modules.bootstrap.forms import BootstrapForm
 from indico.modules.cephalopod.util import register_instance
@@ -37,6 +36,7 @@ from indico.util.string import to_unicode
 from indico.util.system import get_os
 from indico.web.flask.templating import get_template_module
 from indico.web.flask.util import url_for
+from indico.web.rh import RH
 from indico.web.util import url_for_index
