+++ indico/modules/categories/controllers/base.py
@@ -19,9 +19,9 @@ from __future__ import unicode_literals
 from flask import request, session
 from werkzeug.exceptions import Forbidden, NotFound
 
-from indico.legacy.webinterface.rh.base import RH
 from indico.modules.categories.models.categories import Category
 from indico.util.i18n import _
+from indico.web.rh import RH
 
 
 class RHCategoryBase(RH):
