+++ indico/modules/events/registration/api.py
@@ -20,10 +20,10 @@ from flask import jsonify, request
 from sqlalchemy.orm import joinedload
 from werkzeug.exceptions import BadRequest, Forbidden
 
-from indico.legacy.webinterface.rh.base import RH
 from indico.modules.events.models.events import Event
 from indico.modules.events.registration.util import build_registration_api_data, build_registrations_api_data
 from indico.modules.oauth import oauth
+from indico.web.rh import RH
 
 
 class RHAPIRegistrant(RH):
