+++ indico/modules/events/registration/controllers/compat.py
@@ -18,9 +18,9 @@ from __future__ import unicode_literals
 
 from flask import current_app, redirect, request
 
-from indico.legacy.webinterface.rh.base import RHSimple
 from indico.modules.events.registration.models.legacy_mapping import LegacyRegistrationMapping
 from indico.web.flask.util import url_for
+from indico.web.rh import RHSimple
 
 
 @RHSimple.wrap_function
